Support shader print on mac

+++ Source/FrameWork/RenderResource/UniformBuffer.h
@@ -136,8 +136,8 @@ namespace FW
 
 	enum class UniformBufferUsage : uint32
 	{
-		Persistant = GpuBufferUsage::Uniform,
-		Temp = GpuBufferUsage::Uniform | GpuBufferUsage::Temporary,
+		Persistant = (uint32)GpuBufferUsage::Uniform,
+		Temp = (uint32)(GpuBufferUsage::Uniform | GpuBufferUsage::Temporary),
 	};
 	
 	//There is no declared struct for uniform buffer,
